AS in commit 16125 change forgotten type from long to double.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@16152 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/interpolation/CexTraceAnalysisDirection.java
@@ -342,7 +342,7 @@ enum CexTraceAnalysisDirection {
 
       // check for overflow
       if(currentScore < 0) {
-        return Long.MAX_VALUE / varNames.size();
+        return Double.MAX_VALUE / varNames.size();
       }
     }
